+++ src/rt/theater/material/material.hh
@@ -7,31 +7,31 @@ namespace rt {
 class Material {
  public:
   struct Phong {
-    float ka;
-    float kd;
-    float ks;
-    float exp;
+    double ka;
+    double kd;
+    double ks;
+    double exp;
   };
 
   virtual ~Material() noexcept = default;
 
   const Phong& phong() const noexcept;
-  float reflectance() const noexcept;
-  float transmittance() const noexcept;
-  float refraction() const noexcept;
+  double reflectance() const noexcept;
+  double transmittance() const noexcept;
+  double refraction() const noexcept;
 
-  Material& SetPhong(float ka, float kd, float ks, float exp) noexcept;
-  Material& SetReflectance(float value) noexcept;
-  Material& SetTransmittance(float value) noexcept;
-  Material& SetRefraction(float value) noexcept;
+  Material& SetPhong(double ka, double kd, double ks, double exp) noexcept;
+  Material& SetReflectance(double value) noexcept;
+  Material& SetTransmittance(double value) noexcept;
+  Material& SetRefraction(double value) noexcept;
 
   virtual std::string ToString() const = 0;
 
  protected:
   Phong phong_;
-  float reflectance_;
-  float transmittance_;
-  float refraction_;
+  double reflectance_;
+  double transmittance_;
+  double refraction_;
 };
 
 }  // namespace rt
